Add files via upload
The gradient descent algorithm in this code is designed to optimize parameters for a logistic regression model. It initializes weights (w1, w2) and bias (bias) and iterates over a specified number of epochs. In each epoch, it calculates predictions (y_predicted) using a sigmoid activation function, computes the log loss between predicted and true labels (y_true), and determines gradients for weights and bias. These gradients are used to update weights and bias in the direction that minimizes the log loss, guided by a learning rate (learning_rate). The goal is to iteratively improve the model's parameters until convergence, achieving a model that accurately predicts binary outcomes based on input features (age and affordability)